+++ src/Bitpay/Client/Client.php
@@ -555,9 +555,13 @@ class Client implements ClientInterface
     {
         $this->request = $this->createNewRequest();
         $this->request->setMethod(Request::METHOD_GET);
-        $this->request->setPath(sprintf('invoices/%s?token=%s', $invoiceId, $this->token->getToken()));
-        $this->addIdentityHeader($this->request);
-        $this->addSignatureHeader($this->request);
+        if ($this->token->getFacade() === 'merchant') {
+            $this->request->setPath(sprintf('invoices/%s?token=%s', $invoiceId, $this->token->getToken()));
+            $this->addIdentityHeader($this->request);
+            $this->addSignatureHeader($this->request);
+        } else {   
+            $this->request->setPath(sprintf('invoices/%s', $invoiceId));
+        }
         $this->response = $this->sendRequest($this->request);
         $body = json_decode($this->response->getBody(), true);